requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.8
PyMuPDF==1.23.8
PyPDF2==3.0.1
pytesseract==0.3.10
pillow==10.1.0 